AB# 705, AB#706, AB#708, AB#804 opened and resolved. Items now shows description when clicked.
